move visibility helpers to lib/visibility
